+++ mLRS/Common/hal/esp/esp-device_conf.h
@@ -149,7 +149,7 @@
   #define FREQUENCY_BAND_2P4_GHZ
 #endif
 
-#ifdef TX_ELRS_RADIOMASTER_TX16S_2400_ESP32
+#ifdef TX_ELRS_RADIOMASTER_INTERNAL_2400_ESP32
   #define DEVICE_NAME "RM TX16S Int 2.4G"
   #define DEVICE_IS_TRANSMITTER
   #define DEVICE_HAS_SX128x
