+++ mLRS/Common/hal/esp/tx-hal-radiomaster-int-2400-esp32.h
