+++ platformio.ini
@@ -284,7 +284,7 @@ build_flags =
 
 ;-- some ELRS Tx Internal Boards
 
-[env:tx-radiomaster-tx16s-internal-2400]
+[env:tx-radiomaster-internal-2400]
 extends = env_common_tx_esp32
 board = esp32dev
 build_src_filter = 
@@ -292,7 +292,7 @@ build_src_filter =
   +<modules/sx12xx-lib/src/sx128x.cpp>
 build_flags = 
   ${env_common_tx_esp32.build_flags}
-  -D TX_ELRS_RADIOMASTER_TX16S_2400_ESP32
+  -D TX_ELRS_RADIOMASTER_INTERNAL_2400_ESP32
 
 
 [env:tx-radiomaster-bandit-micro-900-sik-telem]
